requirements.txt
aiogram==3.26.0
mysql-connector-python==9.6.0
python-dotenv==1.2.2
